# Re-use work of getting state for a given `state_group` (`_get_state_groups_from_groups`)