--- a/changelog.d/15617.feature
+++ b/changelog.d/15617.feature
@@ -1,0 +1,1 @@
+Make `/messages` faster by efficiently grabbing state out of database whenever we have to backfill and process new events.